+++ src/api_virtual.c
@@ -16,7 +16,7 @@ int initVirtual() { return 0; }
 /**
  * basic temperature simulation
  */
-int getTemperatureVirtual() {
+float getTemperatureVirtual() {
   static int t = 23; /**< starting temperature */
 
   if (heater_status_virtual == HEATER_ON) {
